+++ packages/orq-rc/RELEASES.md
@@ -4849,3 +4849,13 @@ Based on:
 - [typescript v4.0.0-rc.15] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.15 - packages/orq-rc
+
+## 2025-11-07 12:14:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.650.0 (2.743.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.16] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.16] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.16 - packages/orq-rc